ShareProfitTracker_Cloud/backend/main.py
"""
FastAPI Backend for ShareProfitTracker Cloud
Provides authentication and data sync endpoints
"""

from fastapi import FastAPI, HTTPException, Depends, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Import routers
from routers import auth, stocks, sync

app = FastAPI(
    title="ShareProfitTracker Cloud API",
    description="Cloud sync API for ShareProfitTracker desktop application",
    version="1.0.0"
)

# CORS middleware for web dashboard
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure in production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Include routers
app.include_router(auth.router, prefix="/auth", tags=["Authentication"])
app.include_router(stocks.router, prefix="/api/stocks", tags=["Stocks"])
app.include_router(sync.router, prefix="/api/sync", tags=["Sync"])

@app.get("/")
async def root():
    return {
        "message": "ShareProfitTracker Cloud API",
        "version": "1.0.0",
        "status": "running"
    }

@app.get("/health")
async def health_check():
    return {"status": "healthy"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
